+++ app/models/auth-tokens/app-tokens-scopes.ts
@@ -108,6 +108,7 @@ export const appTokensScopes = [
       'GET /vN/calendar/:username/:year/:month/:day',
       'GET /vN/attachments/:attId',
       'GET /vN/attachments/:attId/:type',
+      'POST /vN/attachments/byIds',
     ],
   },
   {