+++ app/routes/api/v2/AttachmentsRoute.js
@@ -8,4 +8,6 @@ export default function addRoutes(app) {
   app.post('/attachments/my/sanitize', controller.mySanitize);
   app.get('/attachments/:attId', controller.getById);
   app.get('/attachments/:attId/:type', controller.getPreview);
+  // We use POST here because this method can accept many post IDs
+  app.post('/attachments/byIds', controller.getByIds);
 }