+++ CHANGELOG.md
@@ -50,6 +50,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [app/serializers/v2/attachment.ts](app/serializers/v2/attachment.ts) file.
 - The new `GET /vN/attachments/:attId` API endpoint returns the attachment by
   its ID.
+- The new `POST /vN/attachments/byIds` API endpoint returns attachments by their
+  IDs. Request body: `{"ids": [...]}`. Response body: `{ "attachments": [...],
+  "users": [...], "idsNotFound": [...] }`.
 - The new `GET /vN/attachments/:attId/:type` API endpoint returns the preview or
   the original of the attachment. The _type_ parameter can be 'original',
   'image', 'video' or 'audio'. The returned data is a JSON object with the following